Changed directory stuff to work like 7zip. Added symlinks for each file in a folder.

+++ CPP/7zip/UI/Console/Explode.cpp
@@ -13,8 +13,11 @@
 #include "../../Common/LimitedStreams.h"
 #include "../../Compress/CopyCoder.h"
 #include "../../../Windows/FileDir.h"
-
-#include <assert.h>
+#ifdef ENV_UNIX
+#include <unistd.h>
+#include <errno.h>
+#include "../../../Common/UTFConvert.h"
+#endif
 
 using namespace NWindows;
 
@@ -37,7 +40,7 @@ static HRESULT WriteRange(IInStream *inStream, ISequentialOutStream *outStream,
 }
 
 // Strip the file name from the full path (if present) and append trailing /
-UString StripFile(UString& path) 
+UString StripFile(UString& path, bool remove=true) 
 {
 	UString file;
 	// remove file name
@@ -46,10 +49,10 @@ UString StripFile(UString& path)
 		int namelen = path.Length() - last;
 
 		file = path.Right(namelen - 1);		
-		path.Delete(last + 1, namelen - 1);
+		if (remove)	path.Delete(last + 1, namelen - 1);
 	} else {// file not directory
 		file = path;
-		path.Empty();
+		if (remove)	path.Empty();
 	}
 	return file;
 }
@@ -60,6 +63,7 @@ UString StripFile(UString& path)
 HRESULT ExplodeArchives(CCodecs *codecs, const CIntVector &formatIndices,
 	bool stdInMode,
 	UStringVector &arcPaths, UStringVector &arcPathsFull,
+	const UString& outputPath,
 	UInt64 &numErrors)
 {
 	int numArcs = arcPaths.Size();
@@ -67,9 +71,11 @@ HRESULT ExplodeArchives(CCodecs *codecs, const CIntVector &formatIndices,
 	{
 		const UString &archivePath = arcPaths[i];
 		
-		UString outputPath = arcPaths[i];
+		/*UString outputPath = arcPaths[i];
 		outputPath.Replace(L'\\', L'/'); // linux and windows consistent
 		const UString archiveName = StripFile(outputPath);
+		outputPath.Empty();*/
+		const UString archiveName = StripFile((UString&)archivePath, false);
 
 		g_StdOut << "Outputting into : " << outputPath << endl;
 
@@ -169,7 +175,7 @@ HRESULT ExplodeArchives(CCodecs *codecs, const CIntVector &formatIndices,
 		CObjectVector<CArchiveDatabase> exploded;
 		CRecordVector<UInt64> folderSizes, folderPositions;
 		szHandler->Explode(exploded, folderSizes, folderPositions);
-
+	
 		if (exploded.Size() == 0) {
 			SHOW_ERROR("Empty archive!");
 			continue;
@@ -224,9 +230,6 @@ HRESULT ExplodeArchives(CCodecs *codecs, const CIntVector &formatIndices,
 			RINOK(WriteRange(inStream, out.SeqStream, 
 			folderStartPackPos, folderLen, NULL));
 
-			// need to set this correctly, the main header isn't being compressed
-			// todo: implement properly, see 7zHandler.h (CHandler)
-			// seems to be fixed now, but broke encapsulation on 7zHandler
 			CCompressionMethodMode method, headerMethod;
 			szHandler->SetCompressionMethod(method, headerMethod);
 
@@ -235,6 +238,31 @@ HRESULT ExplodeArchives(CCodecs *codecs, const CIntVector &formatIndices,
 
 			out.WriteDatabase(exploded[x], &headerMethod, headerOptions);
 			out.Close();
+
+#ifdef ENV_UNIX
+			// Create a symlink for each file in the folder.
+			// This makes it seem as though each file is individually accessible.
+			for (int fileIndex = 0; fileIndex < exploded[x].Files.Size(); fileIndex++) {
+				AString oldfile, newfile;
+				UString woldfile = sstream.str().c_str();
+				UString wnewfile = outputPath + relativeFilePath + exploded[x].Files[fileIndex].Name + L".7z";
+				ConvertUnicodeToUTF8(woldfile, oldfile);
+				ConvertUnicodeToUTF8(wnewfile, newfile);
+				const char* link_to = oldfile.GetBuffer();
+				const char* link_name = newfile.GetBuffer();
+				unlink(link_name);// should ask user
+				//g_StdOut << "Creating symlink to '" << link_to << "' called '" << link_name << "'" << endl;
+				int status = symlink(link_to, link_name);
+				if (status == -1) {
+					AString error = "Couldn't create symlink for '";
+					error += newfile;
+					error += "'";
+					SHOW_ERROR(error);
+					g_StdOut << "Error: " << errno << endl;
+					
+				}
+			}
+#endif
 		}		
 
 		archiveLink.Close(); // not needed but oh well
